requirements.txt
django
djangorestframework
drf_yasg
django-environ
psycopg2-binary
celery
django-celery-beat
redis
aiohttp